test(coverage): drift-refusal tests live beside the verified constructor, cited to the spec
FileExecution::new is the implementation of the spec's Drift response
(trust-taxonomy): a coverage key outside the classified source refuses
construction and the caller reports the conservative Unknown. The
constructor now carries the type=implementation citation, and its
precondition-boundary tests move from the glue (excluded from
annotation scanning as a fixture carrier, issue #226) into
duvet-coverage where the refusal tests carry type=test citations.
Snapshot regenerated: Drift is now cited implementation+test.
